Correct tests, add read/write endpoint
