+++ tests/conftest.py
@@ -1,5 +1,15 @@
 import pytest
 import os
+import requests
+
+
+class MockedResponse:
+    def __init__(self):
+        def jsonFunc():
+            return {"message": "Local API enabled", "apiKey": "fakelocalkey"}
+
+        self.ok = True
+        self.json = jsonFunc
 
 
 def create_fake_cred_file():
@@ -12,10 +22,17 @@ def create_fake_cred_file():
         f.close()
 
 
-def remove_fake_cred_file():
+def create_fake_token_file():
+    with open(os.path.dirname(os.path.dirname(__file__)) + "/local.txt", "w") as f:
+        f.write("fakeLocalKey\n")
+        f.write("10.0.0.2")
+        f.close()
+
+
+def remove_fake_cred_file(fileName):
     try:
         filePath = os.path.dirname(os.path.dirname(__file__))
-        os.remove(filePath + "/credentials.txt")
+        os.remove(filePath + f"/{fileName}.txt")
     except OSError:
         pass
 
@@ -24,9 +41,29 @@ def remove_fake_cred_file():
 def with_credentials():
     create_fake_cred_file()
     yield True
-    remove_fake_cred_file()
+    remove_fake_cred_file("credentials")
 
 
 @pytest.fixture
 def no_credentials():
-    remove_fake_cred_file()
+    remove_fake_cred_file("credentials")
+
+
+@pytest.fixture
+def with_token_file():
+    create_fake_token_file()
+    yield True
+    remove_fake_cred_file("local")
+
+
+@pytest.fixture
+def no_token_file():
+    remove_fake_cred_file("local")
+
+
+@pytest.fixture
+def patched_requests(monkeypatch):
+    def mocked_post(uri, *args, **kwargs):
+        return MockedResponse()
+
+    monkeypatch.setattr(requests, "post", mocked_post)
