+++ vestaboard/vbUrls.py
@@ -1,4 +1,5 @@
-subscription = 'https://platform.vestaboard.com/subscriptions'
+subscription = "https://platform.vestaboard.com/subscriptions"
 post = "https://platform.vestaboard.com/subscriptions/{}/message"
 enableLocal = "http://{}:7000/local-api/enablement"
 postLocal = "http://{}:7000/local-api/message"
+readWrite = "https://rw.vestaboard.com"
